requirements.txt
dj-database-url==0.5.0
Django==3.2.7
django-crispy-forms==1.12.0
django-extensions==3.1.3
psycopg2-binary==2.9.1
python-decouple==3.4
whitenoise==5.3.0
django-import-export==2.6.0